refactor(db): Move statistics read into dedicated class
* Add DbStatsReader class
* Simplify and typefix method names
* Fix error condition with emtpy spawn data
* Add logger to flask route
* Optimize some SQL queries to facilitate index use
* Style SQL
